+++ scripts/ch_lib/model_action_civitai.py
@@ -16,7 +16,7 @@ def get_metadata_skeleton():
     """
     Used to generate at least something when model is not on civitai.
     """
-    return {
+    metadata = {
         "id": "",
         "modelId": "",
         "name": "",
@@ -40,8 +40,12 @@ def get_metadata_skeleton():
                 }
             }
         ],
-        "downloadUrl": ""
+        "tags": [],
+        "downloadUrl": "",
+        "skeleton_file": True
     }
+
+    return metadata
 
 
 def scan_single_model(filename, root, model_type, refetch_old, delay):
@@ -55,38 +59,43 @@ def scan_single_model(filename, root, model_type, refetch_old, delay):
     item = os.path.join(root, filename)
     _, ext = os.path.splitext(item)
     if ext not in model.EXTS:
-        return False
+        yield False
 
     # find a model, get info file
     info_file, sd15_file = model.get_model_info_paths(item)
 
+    output = ""
+
     # check info file
     if model.metadata_needed(info_file, sd15_file, refetch_old):
-        util.printD(f"Creating model info for: {filename}")
+        output = f"Creating model info for: {filename}"
+        util.printD(output)
+        yield output
+
         # get model's sha256
         sha256_hash = util.gen_file_sha256(item)
 
         if not sha256_hash:
             output = f"failed generating SHA256 for model: {filename}"
             util.printD(output)
-            return False
+            yield output
+            yield False
 
         # use this sha256 to get model info from civitai
         model_info = civitai.get_model_info_by_hash(sha256_hash)
 
-        if (model_info == {}) and not model_info.get("id", None):
+        if not model_info:
             model_info = dummy_model_info(item, sha256_hash, model_type)
 
         model.process_model_info(item, model_info, model_type, refetch_old=refetch_old)
 
         # delay before next request, to prevent being treated as a DDoS attack
-        util.printD(f"delay: {delay} second")
         time.sleep(delay)
 
     else:
         util.printD(f"Model metadata not needed for {filename}")
 
-    return True
+    yield True
 
 
 def scan_model(scan_model_types, max_size_preview, nsfw_preview_threshold, refetch_old):
@@ -103,7 +112,8 @@ def scan_model(scan_model_types, max_size_preview, nsfw_preview_threshold, refet
     if not scan_model_types:
         output = "Model Types is None, can not scan."
         util.printD(output)
-        return output
+        yield output
+        return
 
     model_types = []
 
@@ -122,7 +132,14 @@ def scan_model(scan_model_types, max_size_preview, nsfw_preview_threshold, refet
         util.printD(f"Scanning path: {model_folder}")
         for root, _, files in os.walk(model_folder, followlinks=True):
             for filename in files:
-                success = scan_single_model(filename, root, model_type, refetch_old, delay)
+                success = None
+                for result in scan_single_model(filename, root, model_type, refetch_old, delay):
+                    if isinstance(result, str):
+                        yield result
+                        continue
+
+                    success = result
+                    break
 
                 if not success:
                     continue
@@ -176,6 +193,7 @@ def dummy_model_info(path, sha256_hash, model_type):
     # We can't get data on the model from civitai, but some models
     # do store their training data.
     trained_words = model_info["trainedWords"]
+    tags = model_info["tags"]
 
     try:
         file_metadata = sd_models.read_metadata_from_safetensors(path)
@@ -186,9 +204,21 @@ def dummy_model_info(path, sha256_hash, model_type):
 
     tag_frequency = file_metadata.get("ss_tag_frequency", {})
 
-    for tag in tag_frequency.keys():
-        word = re.sub(r"^\d+_", "", tag)
+    for trained_word in tag_frequency.keys():
+        # kohya training scripts use
+        # `{iterations}_{trained_word}`
+        # for training finetune concepts.
+        word = re.sub(r"^\d+_", "", trained_word)
         trained_words.append(word)
+
+        # "tags" in this case are just words used in image captions
+        # when training the finetune model.
+        # They may or may not be useful for prompting
+        for tag in tag_frequency[trained_word].keys():
+            tag = tag.replace(",", "").strip()
+            if tag == "":
+                continue
+            tags.append(tag)
 
     return model_info
 
@@ -325,48 +355,48 @@ def get_model_info_by_url(model_url_or_id:str) -> tuple:
     """
     util.printD(f"Getting model info by: {model_url_or_id}")
 
-    # parse model id
-    model_id = civitai.get_model_id_from_url(model_url_or_id)
-    if not model_id:
-        util.printD("Could not parse model id from url or id")
+    try:
+        # parse model id
+        model_id = civitai.get_model_id_from_url(model_url_or_id)
+
+        # download model info
+        model_info = civitai.get_model_info_by_id(model_id)
+
+        # parse model type, model name, subfolder, version from this model info
+        # get model type
+        civitai_model_type = model_info["type"]
+
+        if civitai_model_type not in civitai.MODEL_TYPES:
+            util.printD(f"This model type is not supported: {civitai_model_type}")
+            return None
+
+        model_type = civitai.MODEL_TYPES[civitai_model_type]
+
+        # get model type
+        model_name = model_info["name"]
+
+        # get version lists
+        model_versions = model_info["modelVersions"]
+
+    except (KeyError, ValueError, TypeError) as e:
+        util.printD(f"An error occurred while attempting to process model info: {e}")
         return None
 
-    model_info = civitai.get_model_info_by_id(model_id)
-    if model_info is None:
-        util.printD("Connection to Civitai API service failed. Wait a while and try again")
-        return None
-
-    if not model_info:
-        util.printD("Failed to get model info from url or id")
-        return None
-
-    # parse model type, model name, subfolder, version from this model info
-    # get model type
-    civitai_model_type = model_info.get("type", None)
-    if civitai_model_type not in civitai.MODEL_TYPES:
-        util.printD(f"This model type is not supported: {civitai_model_type}")
-        return None
-
-    model_type = civitai.MODEL_TYPES[civitai_model_type]
-
-    # get model type
-    model_name = model_info.get("name", None)
-    if model_name is None:
-        util.printD("model name is Empty")
-        model_name = ""
-
-    # get version lists
-    model_versions = model_info.get("modelVersions", None)
-    if model_versions is None:
-        util.printD("modelVersions is Empty")
-        return None
-
+    filenames = []
     version_strs = []
     for version in model_versions:
         # version name can not be used as id
         # version id is not readable
         # so , we use name_id as version string
         version_str = f'{version["name"]}_{version["id"]}'
+
+        filename = ""
+        try:
+            filename = version["files"][0]["name"]
+        except (ValueError, KeyError):
+            pass
+
+        filenames.append(filename)
         version_strs.append(version_str)
 
     # get folder by model type
@@ -384,7 +414,7 @@ def get_model_info_by_url(model_url_or_id:str) -> tuple:
     """)
     util.printD(msg)
 
-    return (model_info, model_name, model_type, subfolders, version_strs)
+    return (model_info, model_name, model_type, filenames, subfolders, version_strs)
 
 
 def get_ver_info_by_ver_str(version_str:str, model_info:dict) -> dict:
@@ -483,11 +513,13 @@ def get_id_and_dl_url_by_version_str(version_str:str, model_info:dict) -> tuple:
     return (version_id, download_url)
 
 
-def download_all(model_folder, version_id, ver_info, duplicate):
+def download_all(filename, model_folder, ver_info, headers, duplicate):
     """
     get all download url from files info
     some model versions have multiple files
     """
+
+    version_id = ver_info["id"]
 
     download_urls = []
 
@@ -506,75 +538,98 @@ def download_all(model_folder, version_id, ver_info, duplicate):
     if result:
         output = "This model version already exists"
         util.printD(output)
-        return (False, output)
+        yield (False, output)
 
     # download
     success = False
     output = ""
     filepath = ""
+    file_candidate = ""
     total = len(download_urls)
+    errors = []
+    errors_count = 0
+
     for count, url in enumerate(download_urls):
+
+        snippet = ""
+        if errors_count > 0:
+            snippet = f"| {errors_count}/{total} models failed"
 
         # webui visible progress bar
         for result in downloader.dl_file(
-            url, folder=model_folder, duplicate=duplicate
+            url, folder=model_folder, duplicate=duplicate,
+            headers=headers
         ):
-            if isinstance(result, str):
-                yield f"{result} | {count}/{total} models"
-                continue
+            if not isinstance(result, str):
+                success, output = result
+                break
 
-            success, output = result
+            yield f"{result} | {count}/{total} models {snippet}"
 
         if not success:
-            downloader.error(url, output)
+            errors.append(downloader.error(url, output))
+            errors_count += 1
             continue
 
-        if url == ver_info["downloadUrl"]:
-            filepath = output
+        file_candidate = output
 
+        if url == ver_info["downloadUrl"]:
+            filepath = file_candidate
 
     if not filepath:
-        filepath = output
+        filepath = file_candidate
 
-    yield (True, filepath)
+    additional = None
+    if errors_count > 0:
+        additional = "\n\n".join(errors)
+
+        if errors_count == total:
+            yield (False, additional)
+            return
+
+    yield (True, filepath, additional)
 
 
-def download_one(model_folder, download_url, duplicate):
+def download_one(filename, model_folder, ver_info, headers, duplicate):
     """
     only download one file
     get download url
     """
 
+    download_url = ver_info["downloadUrl"]
+
     output = ""
-    url = download_url
-    if not url:
-        output = "Failed to get download url, check console log for detail"
+    if not download_url:
+        output = "Failed to find a download url"
         util.printD(output)
-        return None
+        yield (False, output)
 
     # download
     success = False
     for result in downloader.dl_file(
-        url, folder=model_folder, duplicate=duplicate
+        download_url, filename=filename, folder=model_folder,
+        duplicate=duplicate, headers=headers
     ):
-        if isinstance(result, str):
-            yield result
-            continue
+        if not isinstance(result, str):
+            success, output = result
+            break
 
-        success, output = result
+        yield result
 
     if not success:
-        downloader.error(url, output)
-        return None
+        downloader.error(download_url, output)
+        yield (False, output)
 
     yield (True, output)
 
 
 def dl_model_by_input(
-    model_info:dict,
+    ch_state:dict,
     model_type:str,
     subfolder_str:str,
     version_str:str,
+    filename:str,
+    file_ext:str,
     dl_all_bool:bool,
     max_size_preview:bool,
     nsfw_preview_threshold:bool,
@@ -583,12 +638,17 @@ def dl_model_by_input(
     """ download model from civitai by input
         output to markdown log
     """
+
+    model_info = ch_state["model_info"]
+
     if not (model_info and model_type and subfolder_str and version_str):
         output = util.indented_msg(f"""
             Missing Required Parameter in dl_model_by_input. Parameters given:
-            {model_type=}
-            {subfolder_str=}
-            {version_str=}
+            {model_type=}*
+            {subfolder_str=}*
+            {version_str=}*
+            {filename=}
+            {file_ext=}
             {dl_all_bool=}
             {max_size_preview=}
             {nsfw_preview_threshold=}
@@ -597,21 +657,29 @@ def dl_model_by_input(
 
         # Keep model info away from util.indented_msg
         # which can screw with complex strings
-        output = f"{output}\n    {model_info=}"
+        output = f"{output}\n    {model_info=}*\n    * Required"
         util.printD(output)
-        return output
+        yield output
+        return
 
     # get model root folder
     if model_type not in model.folders:
         output = f"Unsupported model type: {model_type}"
         util.printD(output)
-        return output
+        yield output
+        return
 
     folder = ""
     subfolder = ""
-    version_id = ""
     output = ""
     version_info = None
+
+    if filename and file_ext:
+        filename = f"{filename}.{file_ext}"
+    else:
+        # Will force downloader function to get filename
+        # from download headers
+        filename = None
 
     model_root_folder = model.folders[model_type]
 
@@ -628,38 +696,47 @@ def dl_model_by_input(
     if not os.path.isdir(folder):
         output = f"Model folder is not a dir: {folder}"
         util.printD(output)
-        return output
+        yield output
+        return
 
     # get version info
     ver_info = get_ver_info_by_ver_str(version_str, model_info)
     if not ver_info:
         output = "Failed to get version info, check console log for detail"
         util.printD(output)
-        return output
+        yield output
+        return
 
     success = False
-    version_id = ver_info["id"]
+    downloader_fn = download_one
     if dl_all_bool:
-        for result in download_all(folder, version_id, ver_info, duplicate):
-            if isinstance(result, str):
-                yield result
-                continue
+        downloader_fn = download_all
 
-            success, output = result
+    headers = {
+        "content-type": "application/json"
+    }
+    api_key = util.get_opts("ch_civiai_api_key")
+    if api_key:
+        headers["Authorization"] = f"Bearer {api_key}"
 
-    else:
-        for result in download_one(folder, ver_info["downloadUrl"], duplicate):
-            if isinstance(result, str):
-                yield result
-                continue
+    additional = None
+    for result in downloader_fn(filename, folder, ver_info, headers, duplicate):
+        if not isinstance(result, str):
+            if len(result) > 2:
+                success, output, additional = result
+            else:
+                success, output = result
 
-            success, output = result
+            break
+
+        yield result
 
     if not success:
-        return output
+        yield output
+        return
 
     # get version info
-    version_info = civitai.get_version_info_by_version_id(version_id)
+    version_info = civitai.get_version_info_by_version_id(ver_info["id"])
     model.process_model_info(output, version_info, model_type)
 
     # then, get preview image + webui-visible progress
@@ -671,5 +748,7 @@ def dl_model_by_input(
         yield f"Downloading model preview:\n{result}"
 
     output = f"Done. Model downloaded to: {output}"
+    if additional:
+        output = f"{output}. Additionally, the following failures occurred: \n{additional}"
     util.printD(output)
     yield output