+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Change Log
+## 1.7.8
+* Model Downloader now supports changing the filename before downloading.
+* If a download cannot be resumed, it will now be restarted.
+
+## 1.7.7
+* Implemented authentication via API Key.
+  * This allows the user to download models that require authentication.
+
 ## 1.7.6
 * Downloader rewrite:
   * Downloading models will now show download progress in webui.